moved reducers to reducers folder and updated imports
