+++ src/components/Clusters/AddCluster.js
@@ -17,7 +17,7 @@ import Loading from '../common/Loading';
 import Regions from '../../data/Regions';
 import SwipeableList from '../common/SwipeableList';
 import { fetchEksClusters } from '../../reducers/AwsSlice';
-import { addCluster, setCurrentProvider } from './ClustersSlice';
+import { addCluster, setCurrentProvider } from '../../reducers/ClustersSlice';
 
 const AddCluster = ({ navigation }) => {
   const dispatch = useDispatch();
